테트리스 게임 프로젝트 3단계 - 내 풀이

+++ C_Power_Programming/테트리스 게임 프로젝트/blockStageControl.c
@@ -4,11 +4,18 @@
 #include "point.h"
 #include "blockInfo.h"
 #include "keyCurControl.h"
-//#include "blockStageControl.h"
+#include "blockStageControl.h"
+
+#define GBOARD_WIDTH (10)
+#define GBOARD_HEIGHT (20)
+
+#define GBOARD_ORIGIN_X (2*2)
+#define GBOARD_ORIGIN_Y (2)
 
 static int currentBlockModel;
 static int curPosX, curPosY;
 static int rotateSte;
+static int savedBlocks[GBOARD_HEIGHT+1][GBOARD_WIDTH+2] = { 0, };
 
 void InitNewBlockPos(int x, int y) {
 	if (x < 0 || y < 0)
@@ -18,6 +25,16 @@ void InitNewBlockPos(int x, int y) {
 	curPosY = y;
 
 	SetCurrentCursorPos(x, y);
+}
+
+void InitGameBoard(void) {
+	for (int i = 0; i < GBOARD_HEIGHT + 1; i++) {
+		savedBlocks[i][0] = 1;
+		savedBlocks[i][GBOARD_WIDTH + 1] = 1;
+	}
+
+	for (int i = 0; i < GBOARD_WIDTH + 2; i++)
+		savedBlocks[GBOARD_HEIGHT][i] = 1;
 }
 
 void ChooseBlock(void) {
@@ -57,12 +74,19 @@ void DeleteBlock(char blockInfo[][4]) {
 	SetCurrentCursorPos(curPos.x, curPos.y);
 }
 
-void BlockDown(void) {
+int BlockDown(void) {
+	if (DownCrashCheck(blockModel[GetCurrentBlockIdx()])) {
+		BlockPosSaved(blockModel[GetCurrentBlockIdx()]);
+		DrawCurrentState();
+		return 1;
+	}
+
 	DeleteBlock(blockModel[GetCurrentBlockIdx()]);
 	curPosY += 1;
 
 	SetCurrentCursorPos(curPosX, curPosY);
 	ShowBlock(blockModel[GetCurrentBlockIdx()]);
+	return 0;
 }
 
 /*void ArrowMove(void) {
@@ -95,6 +119,9 @@ void BlockDown(void) {
 */
 
 void ShiftLeft(void) {
+	if (LeftCrashCheck(blockModel[GetCurrentBlockIdx()]))
+		return;
+
 	DeleteBlock(blockModel[GetCurrentBlockIdx()]);
 	curPosX -= 2;
 
@@ -103,6 +130,9 @@ void ShiftLeft(void) {
 }
 
 void ShiftRight(void) {
+	if (RightCrashCheck(blockModel[GetCurrentBlockIdx()]))
+		return;
+
 	DeleteBlock(blockModel[GetCurrentBlockIdx()]);
 	curPosX += 2;
 
@@ -113,12 +143,111 @@ void ShiftRight(void) {
 void RotateBlock(void) {
 	int nextRotSte;
 
-	DeleteBlock(blockModel[GetCurrentBlockIdx()]);
-
 	nextRotSte = rotateSte + 1;
 	nextRotSte %= 4;
+
+	if (LeftCrashCheck(blockModel[nextRotSte]) || RightCrashCheck(blockModel[nextRotSte]) || DownCrashCheck(blockModel[nextRotSte]))
+		return;
+
+	DeleteBlock(blockModel[GetCurrentBlockIdx()]);
 	rotateSte = nextRotSte;
 
 	SetCurrentCursorPos(curPosX, curPosY);
 	ShowBlock(blockModel[GetCurrentBlockIdx()]);
+}
+
+void DrawGameBoard(void) {
+	int x, y;
+
+	for (y = 0; y <= GBOARD_HEIGHT; y++) {
+		SetCurrentCursorPos(GBOARD_ORIGIN_X, GBOARD_ORIGIN_Y + y);
+
+		if (y == GBOARD_HEIGHT)
+			printf("¦¦");
+		else
+			printf("¦¢");
+	}
+
+	for (y = 0; y <= GBOARD_HEIGHT; y++) {
+		SetCurrentCursorPos(GBOARD_ORIGIN_X + (GBOARD_WIDTH + 1) * 2, GBOARD_ORIGIN_Y + y);
+
+		if (y == GBOARD_HEIGHT)
+			printf("¦¥");
+		else
+			printf("¦¢");
+	}
+
+	for (x = 1; x < GBOARD_WIDTH + 1; x++) {
+		SetCurrentCursorPos(GBOARD_ORIGIN_X + x * 2, GBOARD_ORIGIN_Y + GBOARD_HEIGHT);
+		printf("¦¡");
+	}
+
+	SetCurrentCursorPos(GBOARD_ORIGIN_X, GBOARD_ORIGIN_Y);
+}
+
+void BlockPosSaved(char blockInfo[][4]) {
+	int y, x;
+	point curPos = GetCurrentCursorPos();
+
+	for (y = 0; y < 4; y++) {
+		for (x = 0; x < 4; x++) {
+			if (blockInfo[y][x] == 1)
+				savedBlocks[curPos.y - GBOARD_ORIGIN_Y + y][(curPos.x - GBOARD_ORIGIN_X)/2 + x] = 1;
+		}
+	}
+}
+
+void DrawCurrentState(void) {
+	int y, x;
+	SetCurrentCursorPos(GBOARD_ORIGIN_X, GBOARD_ORIGIN_Y);
+
+	for (y = 0; y < GBOARD_HEIGHT; y++) {
+		for (x = 0; x < GBOARD_WIDTH; x++) {
+			SetCurrentCursorPos(GBOARD_ORIGIN_X + x * 2, GBOARD_ORIGIN_Y + y);
+			if (savedBlocks[y][x] == 1)
+				printf("¡á");
+		}
+	}
+	SetCurrentCursorPos(GBOARD_ORIGIN_X, GBOARD_ORIGIN_Y);
+}
+
+int LeftCrashCheck(char blockInfo[][4]) {
+	int y, x;
+	point curPos = GetCurrentCursorPos();
+
+	for (y = 0; y < 4; y++) {
+		for (x = 0; x < 4; x++) {
+			if (blockInfo[y][x] == 1 && savedBlocks[curPos.y - GBOARD_ORIGIN_Y + y][(curPos.x - GBOARD_ORIGIN_X)/2 + x - 1] == 1)
+				return 1;
+		}
+	}
+
+	return 0;
+}
+
+int RightCrashCheck(char blockInfo[][4]) {
+	int y, x;
+	point curPos = GetCurrentCursorPos();
+
+	for (y = 0; y < 4; y++) {
+		for (x = 3; x >= 0; x--) {
+			if (blockInfo[y][x] == 1 && savedBlocks[curPos.y - GBOARD_ORIGIN_Y + y][(curPos.x - GBOARD_ORIGIN_X)/2 + x + 1] == 1)
+				return 1;
+		}
+	}
+	return 0;
+}
+
+int DownCrashCheck(char blockInfo[][4]) {
+	int y, x;
+	point curPos = GetCurrentCursorPos();
+
+	for (y = 0; y < 4; y++) {
+		for (x = 0; x < 4; x++) {
+			if (blockInfo[y][x] == 1 && savedBlocks[curPos.y - GBOARD_ORIGIN_Y + y + 1][(curPos.x - GBOARD_ORIGIN_X)/2 + x] == 1)
+				return 1;
+		}
+	}
+
+	return 0;
 }
